# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Heystack - RestAPI",
    "description": "OpenAPI 3.0 specification.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/document": {
      "post": {
        "tags": [
          "document"
        ],
        "summary": "Add a new document",
        "description": "Add a new document to the document",
        "operationId": "addDocument",
        "requestBody": {
          "description": "Create a new document in the document",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "document"
        ],
        "summary": "Get latest 10 document",
        "description": "No required request body",
        "operationId": "getDocument",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      },
      "delete": {
        "tags": [
          "document"
        ],
        "summary": "Delete all document",
        "description": "delete all document",
        "operationId": "deleteDocument",
        "responses": {
          "200": {
            "description": "Gathering were deleted successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Document": {
        "required": [
          "title",
          "slug",
          "url",
          "docUid",
          "description",
          "tags",
          "upvotes",
          "views",
          "readeds",
          "clicks",
          "permissions",
          "meta",
          "payments",
          "allowNotifications",
          "hallOfFame",
          "contentDate",
          "content",
          "spacesArr",
          "archived",
          "ads",
          "allowForks",
          "forks",
          "dateNow",
          "approved"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "name"
          },
          "slug": {
            "type": "string",
            "example": "slug"
          },
          "url": {
            "type": "string",
            "example": "./test.js"
          },
          "docUid": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "description": {
            "type": "string",
            "example": "This is a test"
          },
          "tags": {
            "type": "string",
            "example": "tags"
          },
          "upvotes": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "views": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "readeds": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "clicks": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "permissions": {
            "type": "object"
          },
          "meta": {
            "type": "object"
          },
          "payments": {
            "type": "object"
          },
          "allowNotifications": {
            "type": "boolean",
            "example": true
          },
          "hallOfFame": {
            "type": "object"
          },
          "contentDate": {
            "type": "integer",
            "example": 2022
          },
          "content": {
            "type": "string",
            "example": "content"
          },
          "spacesArr": {
            "type": "string",
            "example": [
              "1",
              "3"
            ]
          },
          "archived": {
            "type": "boolean",
            "example": false
          },
          "ads": {
            "type": "boolean",
            "example": false
          },
          "allowForks": {
            "type": "boolean",
            "example": false
          },
          "forks": {
            "type": "object",
            "example": {
              "original": "",
              "arr": [
                {
                  "emailAddress": "",
                  "id": ""
                }
              ]
            }
          },
          "dateNow": {
            "type": "integer",
            "example": 2022
          },
          "approved": {
            "type": "boolean",
            "example": false
          },
          "hotScore": {
            "type": "integer",
            "example": 100
          }
        }
      },
      "DocumentResponse": {
        "required": [
          "title",
          "slug",
          "url",
          "docUid",
          "description",
          "tags",
          "upvotes",
          "views",
          "readeds",
          "clicks",
          "permissions",
          "meta",
          "payments",
          "allowNotifications",
          "hallOfFame",
          "contentDate",
          "content",
          "spacesArr",
          "archived",
          "ads",
          "allowForks",
          "forks",
          "dateNow",
          "approved"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "name"
          },
          "slug": {
            "type": "string",
            "example": "slug"
          },
          "url": {
            "type": "string",
            "example": "./test.js"
          },
          "docUid": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "description": {
            "type": "string",
            "example": "This is a test"
          },
          "tags": {
            "type": "string",
            "example": "tags"
          },
          "upvotes": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "views": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "readeds": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "clicks": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "permissions": {
            "type": "object"
          },
          "meta": {
            "type": "object"
          },
          "payments": {
            "type": "object"
          },
          "allowNotifications": {
            "type": "boolean",
            "example": true
          },
          "hallOfFame": {
            "type": "object"
          },
          "contentDate": {
            "type": "integer",
            "example": 2022
          },
          "content": {
            "type": "string",
            "example": "content"
          },
          "spacesArr": {
            "type": "string",
            "example": [
              "1",
              "3"
            ]
          },
          "archived": {
            "type": "boolean",
            "example": false
          },
          "ads": {
            "type": "boolean",
            "example": false
          },
          "allowForks": {
            "type": "boolean",
            "example": false
          },
          "forks": {
            "type": "object",
            "example": {
              "original": "",
              "arr": [
                {
                  "emailAddress": "",
                  "id": ""
                }
              ]
            }
          },
          "dateNow": {
            "type": "integer",
            "example": 2022
          },
          "approved": {
            "type": "boolean",
            "example": false
          },
          "hotScore": {
            "type": "integer",
            "example": 100
          },
          "createdAt": {
            "type": "string",
            "example": "2022-08-14T20:16:06.607Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2022-08-14T20:25:52.052Z"
          }
        }
      },
...
    },
...
  }
}